Rename `h.notification.reply_template` to `h.notification.reply`
This module no longer has anything to do with templates, so the old name
doesn't make a lot of sense.
